requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.28.2
botocore==1.31.2
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
coloredlogs==15.0.1
colorful==0.5.5
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.1
distlib==0.3.6
dotmap==1.3.30
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gpustat==1.1
grpcio==1.51.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
inquirer==3.1.3
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
ncnn==1.0.20230517
neptune==1.3.2
networkx==2.8.8
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.8.0.74
openvino-dev==2023.0.1
ordered-set==4.1.0
packaging==23.1
pandas==1.5.2
Pillow==10.0.0
platformdirs==3.9.1
portalocker==2.7.0
prometheus-client==0.17.1
protobuf==4.23.4
psutil==5.9.5
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pyDeprecate==0.3.2
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytz==2023.3
PyYAML==6.0
ray==2.5.1
readchar==4.0.5
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
s3transfer==0.6.1
seaborn==0.12.2
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
swagger-spec-validator==3.0.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
torch==2.0.1
torchmetrics==1.0.0
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.21.0
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
